+++ zcash_client_sqlite/src/wallet/sapling.rs
@@ -9,10 +9,7 @@ use sapling::{self, Diversifier, Nullifier, Rseed};
 use zcash_primitives::{
     consensus::{self, BlockHeight},
     memo::MemoBytes,
-    transaction::{
-        components::{amount::NonNegativeAmount, Amount},
-        TxId,
-    },
+    transaction::{components::Amount, TxId},
     zip32::{AccountId, Scope},
 };
 
@@ -115,7 +112,7 @@ fn to_spendable_note<P: consensus::Parameters>(
         Diversifier(tmp)
     };
 
-    let note_value = NonNegativeAmount::from_nonnegative_i64(row.get(4)?).map_err(|_e| {
+    let note_value: u64 = row.get::<_, i64>(4)?.try_into().map_err(|_e| {
         SqliteClientError::CorruptedData("Note values must be nonnegative".to_string())
     })?;
 
@@ -164,7 +161,7 @@ fn to_spendable_note<P: consensus::Parameters>(
         output_index,
         Note::Sapling(sapling::Note::from_parts(
             recipient,
-            note_value.into(),
+            sapling::value::NoteValue::from_raw(note_value),
             rseed,
         )),
         spending_key_scope,
@@ -484,11 +481,11 @@ pub(crate) mod tests {
     };
     use zcash_primitives::{
         block::BlockHash,
-        consensus::{sapling_zip212_enforcement, BranchId},
+        consensus::BranchId,
         legacy::TransparentAddress,
         memo::{Memo, MemoBytes},
         transaction::{
-            components::{amount::NonNegativeAmount, Amount},
+            components::{amount::NonNegativeAmount, sapling::zip212_enforcement, Amount},
             fees::{
                 fixed::FeeRule as FixedFeeRule, zip317::FeeError as Zip317FeeError, StandardFeeRule,
             },
@@ -1261,7 +1258,7 @@ pub(crate) mod tests {
                 let result = try_sapling_output_recovery(
                     &dfvk.to_ovk(Scope::External),
                     output,
-                    sapling_zip212_enforcement(&st.network(), h1),
+                    zip212_enforcement(&st.network(), h1),
                 );
 
                 if result.is_some() {